requirements.txt
certifi==2018.8.24
Django==2.1.1
django-contrib-comments==1.9.0
django-js-asset==1.1.0
django-latexify==0.2.1
django-mptt==0.9.1
django-taggit==0.23.0
django-threadedcomments==1.2
pytz==2018.5
wincertstore==0.2
